backend/LMS/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-decouple==3.8
twilio==9.4.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
